requirements.txt
anthropic>=0.80
python-dotenv>=1.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
iphone-backup-decrypt>=0.5.0
fpdf2>=2.7
